+++ common/src/main/java/net/bettercombat/config/ServerConfig.java
@@ -5,6 +5,7 @@ import me.shedaniel.autoconfig.annotation.Config;
 import me.shedaniel.cloth.clothconfig.shadowed.blue.endless.jankson.Comment;
 import net.bettercombat.logic.TargetHelper;
 
+import java.util.HashMap;
 import java.util.Map;
 
 @Config(name = "server")
@@ -88,12 +89,13 @@ public class ServerConfig implements ConfigData {
             - `player_relation_to_other`
             (The first relation to be found for the target will be applied.)
             """)
-    public Map<String, TargetHelper.Relation> player_relations = Map.of(
-            "minecraft:player", TargetHelper.Relation.NEUTRAL,
-            "minecraft:villager", TargetHelper.Relation.NEUTRAL,
-            "minecraft:iron_golem", TargetHelper.Relation.NEUTRAL,
-            "guardvillagers:guard", TargetHelper.Relation.NEUTRAL
-    );
+    public HashMap<String, TargetHelper.Relation> player_relations = new HashMap<>() {{
+        put("minecraft:player", TargetHelper.Relation.NEUTRAL);
+        put("minecraft:villager", TargetHelper.Relation.NEUTRAL);
+        put("minecraft:iron_golem", TargetHelper.Relation.NEUTRAL);
+        put("guardvillagers:guard", TargetHelper.Relation.NEUTRAL);
+    }};
+
     @Comment("Relation to unspecified entities those are instance of PassiveEntity(Yarn)")
     public TargetHelper.Relation player_relation_to_passives = TargetHelper.Relation.HOSTILE;
     @Comment("Relation to unspecified entities those are instance of HostileEntity(Yarn)")